refactor: marketplace page refactored
