+++ src/views/MarketplaceView.vue
@@ -1,12 +1,12 @@
 <template>
   <MainLayout>
     <template #content>
-      <ProductGrid />
+      <MarketplaceGrid />
     </template>
   </MainLayout>
 </template>
 
 <script lang="ts" setup>
-import ProductGrid from '@/components/ProductGrid.vue'
+import MarketplaceGrid from '@/components/MarketplaceGrid.vue'
 import MainLayout from '@/layouts/MainLayout.vue'
 </script>
